% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Get data}
\usage{
getdata(campaign_urn, serverurl, token, ...)
}
\arguments{
\item{campaign_urn}{Campaign_urn}

\item{token}{authentication token}

\item{...}{passed on to oh.survey_response.read}

\item{severurl}{Http(s) address of Ohmage server (including /app)}
}
\description{
Wrapper for calling survey_response/read
}

