\name{makeplot}
\alias{makeplot}
\title{Make plot}
\usage{
  makeplot(data, x, y, subset = TRUE, ...)
}
\arguments{
  \item{data}{The dataframe}

  \item{subset}{subset the data. See ?subset}

  \item{...}{passed on to qplot}
}
\description{
  Function to make plot. Currently just wraps around qplot
}

